# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katsumori API documentation",
        "contact": {
            "email": "administrator@katsumori.fun"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:80/api/v1",
            "description": "v1"
        }
    ],
    "paths": {
        "/anime": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Получение списка аниме тайтлов",
                "operationId": "e5bff3bde08d2b68e3ebc66d37f8fd5f",
                "parameters": [
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Сортировка",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "episodes",
                                "mal_id",
                                "mal_score"
                            ],
                            "default": "id"
                        }
                    },
                    {
                        "name": "studios",
                        "in": "query",
                        "description": "Фильтр по студиям (принимает в себя id студий через запятую)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "1,2"
                        }
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "Фильтр по жанрам (принимает в себя id жанров через запятую)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "1,2"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Фильтр по типу тайтла",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "tv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cписка тайтлов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnimeListItem"
                                            }
                                        },
                                        "pagination": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaginationMeta"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaginationLinks"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "post": {
                "tags": [
                    "Anime"
                ],
                "summary": "Добавление нового тайтла",
                "operationId": "785e377dd4f5a8e0a387b454339a933a",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnimeItem"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/anime/{id}": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Получение информации о тайтле по его id",
                "operationId": "122395091efd8c3f960a879b9604bd5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id аниме",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о тайтле",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnimeItem"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "тайтл не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "example": {
                                            "message": "Anime not found",
                                            "errors": {
                                                "Anime": "Anime not found"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Anime"
                ],
                "summary": "Обновление информации о тайтле",
                "description": "Для выполнения данного запроса требуется право доступа anime.update",
                "operationId": "223a2acec636abfe6e4060d8a402a434",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id аниме",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о тайтле",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnimeItem"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "тайтл не найден"
                    },
                    "403": {
                        "description": "недостаточно прав доступа"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/anime/{id}/moderation": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Вывод списка заявок на обновлление информации о тайтле",
                "operationId": "993210e1c826c5c92f122cf9744b5f64",
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HistoryChanges"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/anime/moderation/{id}": {
            "put": {
                "tags": [
                    "Anime"
                ],
                "summary": "Принять заявку на обновление онформации о тайтле по её id",
                "operationId": "516f0b130fe5cd50bd202315941803ea",
                "responses": {
                    "200": {
                        "description": "Успех"
                    },
                    "400": {
                        "description": "Ошибка, вероятнее всего такой заявки не существует"
                    },
                    "403": {
                        "description": "Ошибка доступа"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Anime"
                ],
                "summary": "Отклонить заявку на обновление онформации о тайтле по её id",
                "operationId": "336898dd8f814a1bc93f01a29802a246",
                "responses": {
                    "200": {
                        "description": "Успех"
                    },
                    "400": {
                        "description": "Ошибка, вероятнее всего такой заявки не существует"
                    },
                    "403": {
                        "description": "Ошибка доступа"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnimeItem": {
                "title": "AnimeItem",
                "description": "",
                "properties": {
                    "id": {
                        "description": "Id тайтла",
                        "type": "integer",
                        "example": "1"
                    },
                    "mal_id": {
                        "description": "Id тайтла на MyAnimeList",
                        "type": "integer",
                        "example": "1"
                    },
                    "images": {
                        "$ref": "#/components/schemas/AnimeImages"
                    },
                    "title_jp": {
                        "description": "Название тайтла на японском",
                        "type": "string",
                        "example": "異世界おじさん"
                    },
                    "title_en": {
                        "description": "Название тайтла на английском",
                        "type": "string",
                        "example": "Isekai Ojisan"
                    },
                    "title_ru": {
                        "description": "Название тайтла на русском",
                        "type": "string",
                        "example": "Перерождение Дяди"
                    },
                    "title_synonyms": {
                        "$ref": "#/components/schemas/AnimeTitleSynonyms"
                    },
                    "type": {
                        "type": "string",
                        "example": "tv"
                    },
                    "source": {
                        "type": "string",
                        "example": "manga"
                    },
                    "episodes": {
                        "type": "integer",
                        "example": "10"
                    },
                    "status": {
                        "type": "string",
                        "example": "ongoing"
                    },
                    "airing": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "aired": {
                        "$ref": "#/components/schemas/Aired"
                    },
                    "duration": {
                        "type": "string",
                        "example": "24 min per ep"
                    },
                    "age_rating": {
                        "type": "string",
                        "example": "pg-13"
                    },
                    "mal_score": {
                        "type": "number",
                        "format": "float",
                        "example": "7.93"
                    },
                    "score": {
                        "type": "number",
                        "format": "float"
                    },
                    "statistics": {
                        "type": "number",
                        "format": "float"
                    },
                    "synopsis": {
                        "$ref": "#/components/schemas/Synopsis"
                    },
                    "season": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    },
                    "staff": {
                        "$ref": "#/components/schemas/Staff"
                    },
                    "licensors": {
                        "$ref": "#/components/schemas/Licensor"
                    },
                    "studio": {
                        "$ref": "#/components/schemas/Studio"
                    },
                    "genres": {
                        "$ref": "#/components/schemas/Genre"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    }
                },
                "type": "object"
            },
            "AnimeListItem": {
                "title": "AnimeListItem",
                "description": "",
                "properties": {
                    "id": {
                        "description": "Id тайтла",
                        "type": "integer",
                        "example": "1"
                    },
                    "mal_id": {
                        "description": "Id тайтла на MyAnimeList",
                        "type": "integer",
                        "example": "1"
                    },
                    "images": {
                        "$ref": "#/components/schemas/AnimeImages"
                    },
                    "title_jp": {
                        "description": "Название тайтла на японском",
                        "type": "string",
                        "example": "異世界おじさん"
                    },
                    "title_en": {
                        "description": "Название тайтла на английском",
                        "type": "string",
                        "example": "Isekai Ojisan"
                    },
                    "title_ru": {
                        "description": "Название тайтла на русском",
                        "type": "string",
                        "example": "Перерождение Дяди"
                    },
                    "title_synonyms": {
                        "$ref": "#/components/schemas/AnimeTitleSynonyms"
                    },
                    "type": {
                        "type": "string",
                        "example": "tv"
                    },
                    "source": {
                        "type": "string",
                        "example": "manga"
                    },
                    "episodes": {
                        "type": "integer",
                        "example": "10"
                    },
                    "status": {
                        "type": "string",
                        "example": "ongoing"
                    },
                    "airing": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "aired": {
                        "$ref": "#/components/schemas/Aired"
                    },
                    "duration": {
                        "type": "string",
                        "example": "24 min per ep"
                    },
                    "age_rating": {
                        "type": "string",
                        "example": "pg-13"
                    },
                    "mal_score": {
                        "type": "number",
                        "format": "float",
                        "example": "7.93"
                    },
                    "score": {
                        "type": "number",
                        "format": "float"
                    },
                    "statistics": {
                        "type": "number",
                        "format": "float"
                    },
                    "synopsis": {
                        "$ref": "#/components/schemas/Synopsis"
                    },
                    "season": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    },
                    "producers": {
                        "$ref": "#/components/schemas/Staff"
                    },
                    "licensors": {
                        "$ref": "#/components/schemas/Licensor"
                    },
                    "studio": {
                        "$ref": "#/components/schemas/Studio"
                    },
                    "genres": {
                        "$ref": "#/components/schemas/Genre"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    }
                },
                "type": "object"
            },
...
            "Licensor": {
                "title": "Licensor",
                "description": "Информация о выходе",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "name"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "PaginationLinks",
                "description": "Мета информация для пангинации",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8081/api/v1/anime?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8081/api/v1/anime?page=1"
                    },
                    "prev": {
                        "type": "string",
                        "example": "null"
                    },
                    "next": {
                        "type": "string",
                        "example": "null"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Мета информация для пангинации",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationMetaLinks"
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8081/api/v1/anime"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "to": {
                        "type": "integer",
                        "example": "1"
                    },
                    "total": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}